Add SessionTestHelper.
